backend/requirements.txt
# ── Testing ───────────────────────────────────────────────
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0
httpx==0.27.2                    # Required by FastAPI TestClient

# ── Task queue ────────────────────────────────────────────
celery==5.4.0
redis==5.0.8                     # Redis Python client
flower==2.0.1                    # Celery monitoring dashboard (optional)

# ── Web framework ─────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9          # Required for form data / file uploads

# ── Settings & validation ─────────────────────────────────
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic[email]==2.8.2

# ── Database ──────────────────────────────────────────────
motor==3.5.1                     # Async MongoDB driver
pymongo==4.8.0                   # Sync MongoDB (used by motor internally)
chromadb==0.5.3                  # Local vector database

# ── Auth ──────────────────────────────────────────────────
python-jose[cryptography]==3.3.0 # JWT tokens
passlib[bcrypt]==1.7.4           # Password hashing
bcrypt==4.0.1

# ── OpenAI & LLM ──────────────────────────────────────────
openai==1.45.0
langchain==0.3.1
langchain-openai==0.2.1
langchain-community==0.3.1
langchain-chroma==0.1.3

# ── YouTube processing ────────────────────────────────────
youtube-transcript-api==0.6.2
yt-dlp==2024.8.6                 # Fallback metadata extraction

# ── Text processing ───────────────────────────────────────
tiktoken==0.7.0                  # Token counting (OpenAI tokenizer)
nltk==3.9.1                      # Text cleaning utilities

# ── HTTP & utilities ──────────────────────────────────────
httpx==0.27.2                    # Async HTTP client
python-dotenv==1.0.1
email-validator==2.2.0

# ── Logging ───────────────────────────────────────────────
loguru==0.7.2